Agregar endpoints para Producto

+++ Models/Dtos/ProductoDto.cs
@@ -10,7 +10,7 @@ namespace Models.Dtos
        public String NombreProducto { get; set; }
         public String Categoria { get; set; }
         public String Marca { get; set; }
-        public Double Precio { get; set; } 
-        public Double Costo { get; set; }
+        public decimal Precio { get; set; } 
+        public decimal Costo { get; set; }
     }
 }
